+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- `--config-override=<path>` command line argument. `<path>` is absolute file path
+
 ## [1.15.4] - 2022-12-27
 Happy new years everyone. Here is another minor update that wont change much for most people
 ### Fixed